requirements.txt
streamlit
transformers
torch
torchvision
pillow
opencv-python
tensorflow
huggingface_hub